Add files via upload
These are the files  that I created following The Complete 2020 Web Development Bootcamp by Angela Yu.  

It was a nice project to work on and learned a lot of new concepts such as how Bootstrap works and how to use it to create responsive websites.  Moreover, I learned how to structure HTML and how to use CSS to style it. 

Especially the part about code refactoring was helpful. I learned about better ways to create structure and write code that is easy to understand.